+++ src/test/java/io/javalin/TestContextExtensions.java
@@ -9,7 +9,6 @@ package io.javalin;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import io.javalin.misc.SerializeableObject;
-import io.javalin.util.TestUtil;
 import org.junit.Test;
 import static org.assertj.core.api.Assertions.assertThat;
