+++ src/test/java/io/javalin/TestContextPath.kt
@@ -7,7 +7,6 @@
 package io.javalin
 
 import io.javalin.core.util.Util
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import java.util.function.BiFunction
