+++ src/test/java/io/javalin/TestCookieStore.kt
@@ -7,7 +7,6 @@
 package io.javalin
 
 import io.javalin.rendering.template.TemplateUtil
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import java.util.*
