+++ src/test/java/io/javalin/TestCors.kt
@@ -15,7 +15,6 @@ import io.javalin.core.util.Header.ACCESS_CONTROL_REQUEST_HEADERS
 import io.javalin.core.util.Header.ACCESS_CONTROL_REQUEST_METHOD
 import io.javalin.core.util.Header.ORIGIN
 import io.javalin.core.util.Header.REFERER
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
