+++ src/test/java/io/javalin/TestCustomJetty.kt
@@ -9,7 +9,6 @@ package io.javalin
 
 import com.mashape.unirest.http.Unirest
 import io.javalin.util.TestServlet
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.eclipse.jetty.server.Handler
 import org.eclipse.jetty.server.RequestLog
