+++ src/test/java/io/javalin/TestEncoding.kt
@@ -9,7 +9,6 @@ package io.javalin
 
 import com.mashape.unirest.http.Unirest
 import io.javalin.core.util.Header
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import java.net.URLEncoder
