+++ src/test/java/io/javalin/TestExceptionMapper.kt
@@ -8,7 +8,6 @@
 package io.javalin
 
 import io.javalin.misc.TypedException
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
