+++ src/test/java/io/javalin/TestFilters.kt
@@ -8,7 +8,6 @@
 package io.javalin
 
 import com.mashape.unirest.http.HttpMethod
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
