+++ src/test/java/io/javalin/TestHttpVerbs.kt
@@ -8,8 +8,7 @@
 package io.javalin
 
 import com.mashape.unirest.http.HttpMethod
-import io.javalin.util.TestUtil
-import io.javalin.util.TestUtil.okHandler
+import io.javalin.TestUtil.okHandler
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
