+++ src/test/java/io/javalin/TestLogging.kt
@@ -7,7 +7,6 @@
 package io.javalin
 
 import io.javalin.util.HttpUtil
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import java.util.concurrent.CompletableFuture
