+++ src/test/java/io/javalin/TestMethodNotAllowed.kt
@@ -1,6 +1,5 @@
 package io.javalin
 
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import javax.servlet.http.HttpServletResponse
