+++ src/test/java/io/javalin/TestResponse.kt
@@ -9,7 +9,6 @@ package io.javalin
 
 import com.mashape.unirest.http.HttpMethod
 import io.javalin.core.util.Header
-import io.javalin.util.TestUtil
 import org.apache.commons.io.FileUtils
 import org.apache.commons.io.IOUtils
 import org.assertj.core.api.Assertions.assertThat
