+++ src/test/java/io/javalin/TestStaticDirectorySlash.kt
@@ -8,7 +8,6 @@
 
 package io.javalin
 
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
