+++ src/test/java/io/javalin/TestStaticFiles.kt
@@ -10,7 +10,6 @@ package io.javalin
 import io.javalin.core.util.Header
 import io.javalin.core.util.OptionalDependency
 import io.javalin.staticfiles.Location
-import io.javalin.util.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
